rename bars / Bars --> bar / Bar consistently with trace type
