+++ src/traces/bar/arrays_to_calcdata.js
