+++ src/traces/bar/attributes.js
