+++ src/traces/bar/calc.js
