+++ src/traces/bar/defaults.js
