+++ src/traces/bar/hover.js
