+++ src/traces/bar/index.js
