+++ src/traces/bar/layout_attributes.js
