+++ src/traces/bar/layout_defaults.js
