+++ src/traces/bar/plot.js
