+++ src/traces/bar/set_positions.js
