+++ src/traces/bar/style.js
